fix for save and destruction conflict
ember was throwing an error because a mouse event was triggering a save
at the same time as a module was being destroyed. Added a condition and
listener to defer the destroy operation until the save is completed.

todo: probably shouldn’t save on every mouseup if the coordinates
didn’t change.